Use TogglPy library for api access
